Added Dockerfile - which is extremely basic it's impossible to use a Dockerfile with what we currently have.
Also I've worked on a small main screen for our website
It's just gonna be a "page under construction". Feel free to edit
and experiment with it.

earth.jpg is just the photo used for the html background
(will probably create a separate images folder as we develop our project)

I've also added the database that we'll be using for this project.