This is a class diagram
This class diagram is about at least one user can access the organisation as it is only one organisation, the user can use any additional functions for example to create a username and password to make an account. The exact number of population lives in at least one country, as there are many countries around the world, (and the population changes everyday, people live people die or people migrate to another country to live). One country has one capital city for each country in the world and one capital city in every country contains the exact number of cities(according to the map).